+++ examples/standalone/todomvc-sync-cf/src/livestore/schema.ts
@@ -21,7 +21,7 @@ const app = DbSchema.table(
     newTodoText: DbSchema.text({ default: '' }),
     filter: DbSchema.text({ schema: Filter, default: 'all' }),
   },
-  { deriveMutations: { enabled: true, clientOnly: true } },
+  { deriveMutations: { clientOnly: true } },
 )
 
 export type Todo = DbSchema.FromTable.RowDecoded<typeof todos>